Move the embeddings client under services/embeddings/openai as a provider specific implementation
